requirements.txt
Django
djangorestframework
celery
redis